ci: add test requirements file
