+++ .github/workflows/check.yml
@@ -28,6 +28,7 @@ jobs:
         python -m pip install codecov coverage flake8 pytest pytest-benchmark
         # install dependencies
         pip install -e .[all]
+        pip install tests/requirements.txt
         # show installed packages
         pip freeze
     - name: Benchmark with pytest-benchmark
