+++ plugins/slp_autofocus_fake.py
@@ -1,0 +1,67 @@
+
+import cv2
+import nrefocus
+import qpimage
+
+from shapelink import ShapeLinkPlugin
+
+
+# open and crop hologram image
+img = cv2.imread("hologram_cell_curved_bg.jpg", cv2.IMREAD_GRAYSCALE)
+img_crop = img[300:500, 100:300]
+
+
+class AutofocusFakePlugin(ShapeLinkPlugin):
+    """Displays a rolling mean of a few scalar features"""
+    def __init__(self, *args, **kwargs):
+        super(AutofocusFakePlugin, self).__init__(*args, **kwargs)
+
+    def after_register(self):
+        print(" Preparing for transmission")
+
+    def after_transmission(self):
+        print("\n End of transmission\n")
+
+    def choose_features(self):
+        feats = ["image"]
+        # image will simulate the time take for a hologram
+        return feats
+
+    def handle_event(self, event_data):
+        """Handle a new event"""
+
+        # send image but use the loaded (cropped) hologram
+        # open with qpimage (or correct program)
+        # get phase
+        # use nrefocus to autofocus it
+        # get direction and magnitude of defocus needed
+        # how to pass this value to another program?
+
+        qpi = qpimage.QPImage(data=img_crop, which_data="hologram")
+        field = qpi.field
+
+        # example nrefocus parameters
+        focus_range = (-20, 20)
+        pixel_size = 0.34
+        light_wavelength = 0.532
+
+        afocus = nrefocus.RefocusNumpy(
+            field, light_wavelength, pixel_size)
+        focus_distance = afocus.autofocus(focus_range)
+
+        # send to objective the absolute focus value
+        # this vaue must first be calibrated with the z-stage
+        #  via a log file? Write the focus values to log?
+        # emulate here with print
+        print(focus_distance)
+
+        return False
+
+
+info = {
+    "class": AutofocusFakePlugin,
+    "description": "Autofocus a fake hologram image using "
+                   "nrefocus and qpimage.",
+    "name": "Autofocus",
+    "version": "0.1.1",
+}
