+++ tests/requirements.txt
@@ -1,0 +1,3 @@
+cv2
+nrefocus
+qpimage
